kernelci.cli.base: add APICommand base class
Add an APICommand base class with the command line arguments and a
helper method to get the API object.  This can be used by all CLI
tools that interact with the API.
